+++ src/test/java/com/imsweb/naaccrxml/NaaccrXmlDictionaryUtilsTest.java
@@ -172,6 +172,26 @@ public class NaaccrXmlDictionaryUtilsTest {
             Assert.assertEquals(SpecificationVersion.SPEC_1_5, dictionary.getSpecificationVersion());
             Assert.assertNotNull(dictionary.getDateLastModified());
         }
+
+        // try to read a user dictionary with another error (missing default namespace)
+        exception = false;
+        try (Reader reader = new InputStreamReader(Thread.currentThread().getContextClassLoader().getResourceAsStream("data/dictionary/testing-user-dictionary-220-bad1.xml"))) {
+            NaaccrXmlDictionaryUtils.readDictionary(reader);
+        }
+        catch (IOException e) {
+            exception = true;
+        }
+        Assert.assertTrue(exception);
+
+        // try to read a user dictionary with another error (bad root attribute)
+        exception = false;
+        try (Reader reader = new InputStreamReader(Thread.currentThread().getContextClassLoader().getResourceAsStream("data/dictionary/testing-user-dictionary-220-bad2.xml"))) {
+            NaaccrXmlDictionaryUtils.readDictionary(reader);
+        }
+        catch (IOException e) {
+            exception = true;
+        }
+        Assert.assertTrue(exception);
     }
 
     @Test